+++ src/classical_ciphers/vigenere.rs
@@ -49,7 +49,7 @@ pub fn decrypt_vigenere(ciphertext: &str, key: &str) -> String {
         .collect()
 }
 
-pub fn Menu(PATH: &mut String) -> u8 {
+pub fn Menu(PATH: &mut String) -> usize {
     let mut buf = String::new();
     let mut key = String::new();
     let r;
